Ignore git diffs and change subject max length.
Everything after any line with `diff \-\-git` will be ignored by git, and should be ignored by the linter as well.

Also, github allows subject messages of 68 characters before truncating them, so this is another useful error. Changed the original 50 character limit to a warning and added the 68 character as the new hard limit that will be useful for most people.